Some cleanup / modernization / os x portability from plane. Also start draft big interval arithmetic
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6588 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/geom/hilbert-curve_test.cc
@@ -67,8 +67,8 @@ static void Brightness() {
 // small bases.
 static void Inv() {
   for (int base = 1; base < 11; base++) {
-    const int side = 1 << base;
-    for (int d = 0; d < side * side; d++) {
+    const uint32_t side = 1 << base;
+    for (uint64_t d = 0; d < side * side; d++) {
       const auto [x, y] = HilbertCurve::To2D(base, d);
       CHECK(x >= 0 && x < side);
       CHECK(y >= 0 && y < side);
